requirements.txt
accelerate==1.2.1
torch==2.5.1 --index-url https://download.pytorch.org/whl/cu121
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anykeystore==0.2
async-timeout==5.0.1
asyncio==3.4.3
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
coloredlogs==15.0.1
cryptacular==1.6.2
datasets==3.2.0
defusedxml==0.7.1
dill==0.3.8
editables==0.5
einops==0.8.0
enscons==0.30.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
greenlet==3.1.1
huggingface-hub==0.27.1
humanfriendly==10.0
hupper==1.12.1
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==1.26.3
oauthlib==3.2.2
omegaconf==2.3.0
optimum==1.23.3
packaging==24.2
pandas==2.2.3
PasteDeploy==3.1.0
pbkdf2==1.3
peft==0.14.0
pillow==10.2.0
pip_search==0.0.13
plaster==1.1.2
plaster-pastedeploy==1.0.1
propcache==0.2.1
protobuf==3.20.0
psutil==6.1.1
PuLP==2.9.0
pyarrow==18.1.0
pybind11==2.13.6
pydantic==2.10.4
pydantic_core==2.27.2
pydot==3.0.4
Pygments==2.18.0
pyparsing==3.2.1
pyramid==2.0.2
pyramid-mailer==0.15.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytoml==0.1.21
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
repoze.sendmail==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
safetensors==0.5.2
SCons==4.8.1
sentencepiece==0.2.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.36
sympy==1.13.1
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
transaction==5.0
transformer_engine==1.13.0
transformer_engine_cu12==1.13.0
transformers==4.48.0
translationstring==1.4
triton==3.1.0
trl==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
velruse==1.1.1
venusian==3.1.1
WebOb==1.8.9
WTForms==3.2.1
wtforms-recaptcha==0.3.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zope.deprecation==5.0
zope.interface==7.2
zope.sqlalchemy==3.1
